migraciones

+++ database/migrations/2024_05_14_083149_create_´producto_carrito_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->id(); 
             $table->integer('cantidad');
             
-            $table->integer('producto_id');
-            $table->bigInteger('carrito_id');
+            $table->unsignedBigInteger('producto_id');
+            $table->unsignedBigInteger('carrito_id');
 
             $table->foreign('producto_id')->references('id')->on('_productos');
             $table->foreign('carrito_id')->references('id')->on('carritos');
